class/阮耀文的python实验/设备管理/requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
bcrypt==4.0.1
python-dotenv==1.0.0
cryptography==41.0.7
pyinstaller==6.13.0 